Implemented part of button remapping system
You can assign keys to different descriptive names in the Options Menu, although they don't affect anything in-game yet. Unity's input system is much more painful to work with than it should be, so the only way to really check the last key that was pressed is to get the entire KeyCode enumeration (over 300 values) and check each of them in a huge foreach loop. It's ugly. Next task is to add other options.