Ability to load a level from the level browser; tooltips on UI elements.
When you click a level entry on the level browser, that level is loaded back up in the Level Editor. Likewise, levels saved in the Level Editor all appear in the Level Browser in reverse chronological order. UI items on the toolbar and palette of the Level Editor have descriptive tooltips.