Level Browser shows thumbnails of levels. All tools greyed out where applicable.
When the Save button is pressed, a camera is positioned at the bottom of the level and takes a screenshot of size (800, 200). This image is saved along with the level, and is loaded up when the Level Browser loads the level. The Fill, Rectangle (Filled) and Rectangle (Hollow) tools are greyed out when using a tool larger than (1, 1).