Abstracting tile drawing functions.
About 50% of the way through abstracting the functionality to be agnostic wrt whether a mobile player or the desktop player is drawing, with a wrapper (TileDrawWrapper) around the functionality pumping function calls to TileDraw objects. Undo/redo stacks also abstracted out (TileHistory) but this will likely take more work, so it might be un-abstracted in the near future. Also, added a chocolate bar. Currently nothing will work because it's halfway through the overhaul. But there's chocolate now!