Added antialiasing filters.
UI buttons especially look crisper with SMAA (subpixel morphological antialiasing), although this is an effect found on the Unity Asset Store. The main scene camera and UI camera are both supersampled to retain more detail (this especially works well on tiles with small details, such as food). This is original material.