Implemented event structure between Palette and Toolbar.
LevelEditor will become a Service Locator for Palette and Toolbar; theoretically, this means either of those classes could become abstract and be switched at runtime for different concrete representations and the program would still work. These services use dependency injection to register on the LevelEditor in Awake() and all location of these services is performed in Start() to avoid temporal coupling. LevelEditor uses a neat trick using a static constructor to ensure that a Singleton instance is always created right before the first time one is requested. While this is slow at runtime, in this case it will be requested in Awake() anyway. All ToolbarButtons have been set up in the Leveleditor scene.