Added level naming UI.
Begun the backend functionality for saving and loading a level. A static class, LevelManagement, will be used to persist a level ID across scenes. This ID is used by the internal level database to retrieve the correct level data. An ID of -1 indicates a level that has not been saved before. A level will have local IDs for local storage and global IDs if it has been uploaded online. For online purposes, a version ID will also be used so that users have access to the latest version of a level. Global IDs will be a combination of local ID and user ID. Two local databases will be maintained: local levels created by this user, and global levels downloaded from the Internet that have been stored locally.