Level saving to local database works properly.
When a level ID of -1 is specified, this means it is a new level. The code automatically works out what the level ID should be once saved and saves the level data under (/levels/local/[level ID].dat). If the level ID was specified, it saves data under the specified ID.